Add test expected fail

+++ pointer/pointer.go
@@ -7,6 +7,6 @@ func (w Wallet) Deposit(amount int) {
 
 }
 
-func (w Wallet) Balance() int64 {
+func (w Wallet) Balance() int {
 	return 0
 }
